Move makeSureValueIsWithinLimit() to label.ts
